+++ vehicle/OVMS.V3/components/vehicle_hyundai_ioniqvfl/src/vehicle_hyundai_ioniqvfl.h
@@ -42,7 +42,7 @@ class OvmsVehicleHyundaiVFL : public OvmsVehicle
 
   protected:
     void PollerStateTicker();
-    void IncomingPollReply(canbus* bus, uint16_t type, uint16_t pid, uint8_t* data, uint8_t length, uint16_t mlremain);
+    void IncomingPollReply(canbus* bus, uint16_t type, uint32_t pid, uint8_t* data, uint8_t length, uint16_t mlremain);
 
   protected:
     std::string         m_rxbuf;
