+++ vehicle/OVMS.V3/components/vehicle_jaguaripace/src/vehicle_jaguaripace.h
@@ -39,7 +39,7 @@ class OvmsVehicleJaguarIpace : public OvmsVehicle {
     ~OvmsVehicleJaguarIpace();
 
   protected:
-    void IncomingPollReply(canbus* bus, uint16_t type, uint16_t pid, uint8_t* data, uint8_t length, uint16_t mlremain);
+    void IncomingPollReply(canbus* bus, uint16_t type, uint32_t pid, uint8_t* data, uint8_t length, uint16_t mlremain);
     void IncomingFrameCan1(CAN_frame_t* p_frame) override;
     //void IncomingFrameCan2(CAN_frame_t* p_frame) override;
     //void IncomingFrameCan3(CAN_frame_t* p_frame) override;
@@ -59,13 +59,13 @@ class OvmsVehicleJaguarIpace : public OvmsVehicle {
 
   private:
     void IncomingPollFrame(CAN_frame_t* frame);
-    void IncomingBecmPoll(uint16_t pid, uint8_t* data, uint8_t length, uint16_t remain);
-    void IncomingHvacPoll(uint16_t pid, uint8_t* data, uint8_t length, uint16_t remain);
-    void IncomingBcmPoll(uint16_t pid, uint8_t* data, uint8_t length, uint16_t remain);
-    void IncomingTpmsPoll(uint16_t pid, uint8_t* data, uint8_t length, uint16_t remain);
-    void IncomingTcuPoll(uint16_t pid, uint8_t* data, uint8_t length, uint16_t remain);
+    void IncomingBecmPoll(uint32_t pid, uint8_t* data, uint8_t length, uint16_t remain);
+    void IncomingHvacPoll(uint32_t pid, uint8_t* data, uint8_t length, uint16_t remain);
+    void IncomingBcmPoll(uint32_t pid, uint8_t* data, uint8_t length, uint16_t remain);
+    void IncomingTpmsPoll(uint32_t pid, uint8_t* data, uint8_t length, uint16_t remain);
+    void IncomingTcuPoll(uint32_t pid, uint8_t* data, uint8_t length, uint16_t remain);
 
-    bool SendPollMessage(canbus* bus, uint16_t id, uint8_t type, uint16_t pid);
+    bool SendPollMessage(canbus* bus, uint16_t id, uint8_t type, uint32_t pid);
 
 };
