+++ vehicle/OVMS.V3/components/vehicle_smarteq/src/eq_can_poll.cpp
@@ -77,7 +77,7 @@ static const char *TAG = "v-smarteq";
 /**
  * Incoming poll reply messages
  */
-void OvmsVehicleSmartEQ::IncomingPollReply(canbus* bus, uint16_t type, uint16_t pid, uint8_t* data, uint8_t length, uint16_t remain) {
+void OvmsVehicleSmartEQ::IncomingPollReply(canbus* bus, uint16_t type, uint32_t pid, uint8_t* data, uint8_t length, uint16_t remain) {
   static int last_pid = -1;
   static int last_remain = -1;
   static uint8_t buf[MAX_POLL_DATA_LEN];
