+++ vehicle/OVMS.V3/components/vehicle_vweup/src/vehicle_vweup.h
@@ -295,7 +295,7 @@ protected:
 protected:
   void PollSetState(uint8_t state);
   void PollerStateTicker();
-  void IncomingPollReply(canbus *bus, uint16_t type, uint16_t pid, uint8_t *data, uint8_t length, uint16_t mlremain);
+  void IncomingPollReply(canbus *bus, uint16_t type, uint32_t pid, uint8_t *data, uint8_t length, uint16_t mlremain);
 
 protected:
   void UpdateChargePower(float power_kw);
